fix: table foreign and seeder

+++ ow-interactive/database/migrations/2021_05_22_142548_create_moviments_table.php
@@ -15,8 +15,8 @@ class CreateMovimentsTable extends Migration
     {
         Schema::create('moviments', function (Blueprint $table) {
             $table->id();
-            $table->integer('financial_id')->unsigned();
-            $table->integer('moviment_type_id')->unsigned();
+            $table->bigInteger('financial_id')->unsigned();
+            $table->bigInteger('moviment_type_id')->unsigned();
             $table->double('value', 12, 2);
 
             $table->timestamps();
